pyproject.toml
[tool.poetry]
name = "drf_shoestore"
version = "0.1.0"
description = ""
authors = ["dbmiddle <dbmiddle39@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.7"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.3"
autopep8 = "^1.5.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
